+++ sandbox_std/fibonacci/src/iterator.rs
@@ -2,7 +2,7 @@
 //! Подсчёта числа Фибоначчи через fold
 //!
 //! ```no_run
-//! pub fn iterator_fold(n: u64) -> u64 {
+//! pub fn iterator_fold(n: u8) -> u64 {
 //!     match n {
 //!         0 | 1 => 1,
 //!         n => (0..n).fold((0, 1), |(a, b), _| (b, a + b)).1,
@@ -25,7 +25,7 @@
 /// assert_eq!(iterator_fold(11),144);
 /// assert_eq!(iterator_fold(22),28657);
 /// ```
-pub fn iterator_fold(n: u64) -> u64 {
+pub fn iterator_fold(n: u8) -> u64 {
     match n {
         0 | 1 => 1,
         n => (0..n).fold((0, 1), |(a, b), _| (b, a + b)).1,
